introduce operator NotEqual for CompareFunctionCondition
git-svn-id: https://svn.exedio.com/cope/trunk/persistence@6829 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/testsrc/com/exedio/cope/FieldDoubleTest.java
@@ -33,7 +33,7 @@ public class FieldDoubleTest extends FieldTest
 		assertEquals(null, item.getSomeDouble());
 		assertContains(item, item2, item.TYPE.search(item.someDouble.equal((Double)null)));
 		assertContains(item, item2, item.TYPE.search(item.someDouble.isNull()));
-		assertContains(item.TYPE.search(item.someDouble.notEqual(null)));
+		assertContains(item.TYPE.search(item.someDouble.notEqual((Double)null)));
 		assertContains(item.TYPE.search(item.someDouble.isNotNull()));
 
 		item.someDouble.set(item, new Double(44.44));
@@ -65,7 +65,7 @@ public class FieldDoubleTest extends FieldTest
 			item.TYPE.search(item.someDouble.notEqual(22.22)));
 		assertEquals(list(item2), item.TYPE.search(item.someDouble.equal((Double)null)));
 		assertEquals(list(item2), item.TYPE.search(item.someDouble.isNull()));
-		assertEquals(list(item), item.TYPE.search(item.someDouble.notEqual(null)));
+		assertEquals(list(item), item.TYPE.search(item.someDouble.notEqual((Double)null)));
 		assertEquals(list(item), item.TYPE.search(item.someDouble.isNotNull()));
 
 		assertContains(new Double(22.22), null, search(item.someDouble));
